Inset icon in collection headers, minor ContentEditable refactor
